implements envelope construction in synth class. Tests envelope construction by plotting it in slider notebook
